Move graphical element to creative-coding component with fixed background; make cards partially translucent so the background remains visible and adjust card colors slightly
