+++ src/app/creative-coding/creative-coding.component.html
@@ -1,1 +1,5 @@
-<p>creative-coding works!</p>
+<section class="min-h-screen bg-slate-800 text-white flex items-center justify-center px-6 relative overflow-hidden"
+  appMouseTrail>
+  <!-- container for Three.js canvas -->
+  <div #threeJsCanvas class="w-full h-full absolute top-0 left-0 z-0 pointer-events-none"></div>
+</section>
