+++ src/app/home/home.component.html
@@ -1,8 +1,7 @@
-<section class="min-h-screen bg-slate-800 text-white flex items-center justify-center px-6 relative  overflow-hidden"
-  appMouseTrail>
+<section class="min-h-screen text-white flex items-center justify-center px-6 relative  overflow-hidden" appMouseTrail>
   <!-- canvas for Three.js 3D scene render -->
-  <div #threeJsCanvas class="w-full h-full absolute top-0 left-0 z-0 pointer-events-none">
-  </div>
+  <!--<div #threeJsCanvas class="w-full h-full absolute top-0 left-0 z-0 pointer-events-none">
+  </div> -->
 
   <!-- content remains above the globe -->
   <div class="max-w-6xl text-center relative z-20">
